added for loop for answers
Repeat an event until some condition is met
Very powerful for counting or iterating
One type of loop is called a “For Loop”
Loop a set number of times

	for(int i = 0; i < n; i++)
	{
		//do stuff
}

	int i = 0 parts is runs once before the code block is executed
		sets up the iterator
i < n part is defines the loop condition, tells the loop when to stop, be careful of infinite loops.
i++ part is executes at the end of every loop, used to increment or decrement the iterator.